refactor(imports): ♻️ Replace external @zayne-labs imports with local components - Migrate getElementList from @zayne-labs/ui-react to local for.tsx in   app/(dashboard)/patient/appointment/page.tsx   app/(primary)/daily-tips/DailyTipCard.tsx
feat(dropzone): 🖼️ Implement enhanced file preview handling in DropZoneInput.tsx
- Add PDF/document type detection with dedicated icons
- Implement trash button for file removal
- Improve image preview grid layout

chore(cleanup): 🧹 Remove deprecated primary layout components
- Delete NavBar, Footer, Main and index.ts from app/(primary)/_components
