+++ app/(primary)/-components/AccordionComponent.tsx
@@ -1,9 +1,9 @@
 "use client";
 // FIXME - Change accordions component to another suitable one and then remove use client from this page`
 
+import { getElementList } from "@/components/common/for";
 import { ChevronDownIcon } from "@/components/icons";
 import { Accordion } from "@/components/ui";
-import { getElementList } from "@zayne-labs/ui-react/common/for";
 
 const FAQs = [
 	{
