+++ app/(primary)/-components/NavBar.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { Logo, NavLink } from "@/components/common";
+import { getElementList } from "@/components/common/for";
 import { HamburgerIcon, SearchIcon, XIcon } from "@/components/icons";
 import { Button } from "@/components/ui";
 import { cnMerge } from "@/lib/utils/cn";
 import { useToggle } from "@zayne-labs/toolkit-react";
-import { getElementList } from "@zayne-labs/ui-react/common/for";
 
 function NavBar() {
 	const [isNavShow, toggleNavShow] = useToggle(false);
