+++ app/(primary)/daily-tips/DailyTipCard.tsx
@@ -1,11 +1,11 @@
 "use client";
 
 import { IconBox, NavLink } from "@/components/common";
+import { getElementList } from "@/components/common/for";
 import { Card } from "@/components/ui";
 import type { callBackendApi } from "@/lib/api/callBackendApi";
 import type { TipsResponse } from "@/lib/api/callBackendApi/types";
 import { cnMerge } from "@/lib/utils/cn";
-import { getElementList } from "@zayne-labs/ui-react/common/for";
 import { useDragScroll } from "@zayne-labs/ui-react/ui/drag-scroll";
 import Image from "next/image";
