+++ app/(primary)/daily-tips/[id]/page.tsx
@@ -1,8 +1,8 @@
-import { Main } from "@/app/(primary)/_components";
-import { Await } from "@/components/common";
+import { Main } from "@/app/(primary)/-components";
+import { AwaitRoot } from "@/components/common/await";
+import { getElementList } from "@/components/common/for";
 import { type SingleTip, callBackendApi } from "@/lib/api/callBackendApi";
 import { getTipsResponse } from "@/lib/api/callBackendApi/utils";
-import { getElementList } from "@zayne-labs/ui-react/common/for";
 import Image from "next/image";
 import { ScrollableTipCards } from "../DailyTipCard";
 import HealthFinderLogo from "../HealthFinderLogo";
@@ -68,9 +68,9 @@ async function TipExpandedPage(props: { params: Promise<{ id: string }> }) {
 					Checkout Other Tips
 				</h2>
 
-				<Await.Root promise={tipsResponsePromise} asChild={true}>
+				<AwaitRoot promise={tipsResponsePromise} asChild={true}>
 					<ScrollableTipCards />
-				</Await.Root>
+				</AwaitRoot>
 			</section>
 		</Main>
 	);
