+++ app/(primary)/library/page.tsx
@@ -1,4 +1,4 @@
-import { Main } from "@/app/(primary)/_components";
+import { Main } from "@/app/(primary)/-components";
 import { NavLink } from "@/components/common";
 import { callBackendApi } from "@/lib/api/callBackendApi";
 import type { DiseasesResponse } from "@/lib/api/callBackendApi/types";
