+++ components/common/Await.tsx
@@ -1,5 +1,4 @@
 "use client";
 
-/* eslint-disable react-refresh/only-export-components */
+// eslint-disable-next-line react-refresh/only-export-components
 export * from "@zayne-labs/ui-react/common/await";
-/* eslint-enable react-refresh/only-export-components */
