+++ components/common/for.tsx
@@ -1,4 +1,2 @@
-"use client";
-
 // eslint-disable-next-line react-refresh/only-export-components
-export * from "@zayne-labs/ui-react/common/show";
+export * from "@zayne-labs/ui-react/common/for";
