+++ components/common/index.ts
@@ -1,9 +1,9 @@
 export { default as Logo } from "./Logo/logo";
 export { IconBox } from "./IconBox";
 export { default as SonnerToaster } from "./Toaster";
-export { Show } from "./Show";
-export { Await } from "./Await";
-export { Switch } from "./Switch";
+export { Show } from "./show";
+export { Await } from "./await";
+export { Switch } from "./switch";
 export { default as Overlay } from "./Overlay";
 export { default as NavLink } from "./NavLink";
 export * from "./DropZoneInput";
